+++ src/routes.js
@@ -1,5 +1,6 @@
 const routes = require('express').Router();
 const ArticleController = require('./controllers/ArticleController')
+const UserController = require('./controllers/UserController')
 
 routes
     // CRUD - Articles
@@ -7,5 +8,10 @@ routes
     .post('/articles', ArticleController.create)
     .put('/articles', ArticleController.edit)
     .delete('/articles/:id', ArticleController.remove)
+    // CRUD - Users
+    .get('/users', UserController.index)
+    .post('/users', UserController.create)
+    .put('/users', UserController.edit)
+    .delete('/users/:id', UserController.remove)
 
 module.exports = routes;